% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/evaluate_likelihood_of_evidence_given_hypothesis.R
\name{evaluate_likelihood_ratio_of_evidence_given_hypothesis_set}
\alias{evaluate_likelihood_ratio_of_evidence_given_hypothesis_set}
\title{Evaluates the likelihood ratio of evidence give hypothesis vs evidence given negation of hypothesis.}
\usage{
evaluate_likelihood_ratio_of_evidence_given_hypothesis_set(
  x,
  hypotheses,
  target_variable,
  target_value
)
}
\arguments{
\item{x}{A dataframe.}

\item{hypotheses}{A dataframe containing a set of hypotheses returned by either mre::create_first_set_of_candidate_hypotheses() or mre::create_expanded_set_of_candidate_hypotheses()}

\item{target_variable}{A string representing the variable you are trying to explain.}

\item{target_value}{string representing the particular state of the target variable you are trying to explain.}
}
\value{
A table containing the supplied hypothesis set and their corresponding GBFs determined by the likelihood ratio of seeing the evidence given the hypothesis vs not seeing the hypothesis.
}
\description{
Evaluates the likelihood ratio of evidence give hypothesis vs evidence given negation of hypothesis.
}
\author{
Davide Lorino
}
