% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood_ratio_of_evidence_given_hypothesis.R
\name{likelihood_ratio_of_evidence_given_hypothesis}
\alias{likelihood_ratio_of_evidence_given_hypothesis}
\title{Produces the necessary filters to be evaluated in order to calculate the likelihood ratio of a hypothesis given some evidence.}
\usage{
likelihood_ratio_of_evidence_given_hypothesis(
  x,
  hypothesis,
  target_variable,
  target_value,
  method = "hbs"
)
}
\arguments{
\item{x}{A dataframe.}

\item{hypothesis}{A string representing a given hypothesis e.g. '(age == "old")'.}

\item{target_variable}{A string representing the variable you are trying to explain.}

\item{target_value}{A string representing the state of the variable you are trying to explain.}

\item{method}{A string representing the search algorithm being used - currently only Hierarchical Beam Search is implemented.}
}
\value{
A list containing five filters to be evaluated downstream; the hypothesis, the probability of the evidence given the hypothesis, and the probability of the evidence given the negation of the hypothesis.
}
\description{
Produces the necessary filters to be evaluated in order to calculate the likelihood ratio of a hypothesis given some evidence.
}
\author{
Davide Lorino
}
