% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_odds_of_hypothesis_given_evidence.R
\name{posterior_odds_of_hypothesis_given_evidence}
\alias{posterior_odds_of_hypothesis_given_evidence}
\title{Produces the necessary filters to be evaluated in order to calculate the posterior odds of a hypothesis given evidence.}
\usage{
posterior_odds_of_hypothesis_given_evidence(
  x,
  hypothesis,
  target_variable,
  target_value,
  method = "hbs"
)
}
\arguments{
\item{x}{A dataframe.}

\item{hypothesis}{A string representing a given hypothesis e.g. '(age == "old")'.}

\item{target_variable}{A string representing the variable you are trying to explain.}

\item{target_value}{A string representing the state of the variable you are trying to explain.}

\item{method}{A string representing the search algorithm being used - currently only Hierarchical Beam Search is implemented.}
}
\value{
A list containing five filters to be evaluated downstream; the hypothesis, the conditional probability of hypothesis given evidence, the conditional probability of the negation of the hypothesis given evidence, the marginal probability of the hypothesis, and the marginal probability of the negation of the hypothesis.
}
\description{
Produces the necessary filters to be evaluated in order to calculate the posterior odds of a hypothesis given evidence.
}
\author{
Davide Lorino
}
